add Journal abi encodable struct
