+++ core/src/consensus_state.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::{Checkpoint, Link, ensure};
+use crate::{Checkpoint, Link, abi, ensure};
 use alloy_sol_types::{SolType, SolValue};
 use safe_arith::{ArithError, SafeArith};
 use thiserror::Error;
@@ -45,66 +45,6 @@ pub enum ConsensusError {
 impl From<ArithError> for ConsensusError {
     fn from(err: ArithError) -> Self {
         ConsensusError::ArithError(err)
-    }
-}
-
-/// Private module for ABI encoding and decoding, mapping Rust types to Solidity types.
-mod abi {
-    use crate::{ConsensusError, ensure};
-
-    // Defines the Solidity struct layouts for Checkpoint and State.
-    alloy_sol_types::sol! {
-        struct Checkpoint {
-            uint64 epoch;
-            bytes32 root;
-        }
-
-        struct State {
-            Checkpoint current_justified;
-            Checkpoint finalized;
-        }
-    }
-
-    // ABI conversions for Checkpoint.
-    impl From<&crate::Checkpoint> for Checkpoint {
-        fn from(value: &crate::Checkpoint) -> Self {
-            Self {
-                epoch: value.0.epoch.as_u64(),
-                root: value.0.root,
-            }
-        }
-    }
-
-    impl From<Checkpoint> for crate::Checkpoint {
-        fn from(checkpoint: Checkpoint) -> Self {
-            Self(beacon_types::Checkpoint {
-                epoch: crate::Epoch::new(checkpoint.epoch),
-                root: checkpoint.root,
-            })
-        }
-    }
-
-    // ABI conversions for ConsensusState.
-    impl From<&crate::ConsensusState> for State {
-        fn from(value: &crate::ConsensusState) -> Self {
-            Self {
-                current_justified: (&value.current_justified_checkpoint).into(),
-                finalized: (&value.finalized_checkpoint).into(),
-            }
-        }
-    }
-
-    impl TryFrom<State> for crate::ConsensusState {
-        type Error = ConsensusError;
-
-        fn try_from(state: State) -> Result<Self, Self::Error> {
-            let state = Self {
-                current_justified_checkpoint: state.current_justified.into(),
-                finalized_checkpoint: state.finalized.into(),
-            };
-            ensure!(state.is_valid(), ConsensusError::InvalidState);
-            Ok(state)
-        }
     }
 }
