+++ methods/guest/Cargo.toml
@@ -10,6 +10,7 @@ codegen-units = 1
 lto = "fat"
 
 [dependencies]
+alloy-sol-types = { version = "1.2.1" }
 bincode = { version = "1.3" }
 blst = { version = "0.3.14", default-features = false, features = ["no-threads"] }
 chainspec = { path = "../../chainspec", features = ["mainnet", "sepolia"] }
